withOptionalNode, exampleTypes README, and tons of cleanup
