+++ src/ModalFilterAdder.js
@@ -1,7 +1,9 @@
+import _ from 'lodash/fp'
 import FilterAdder from './FilterAdder'
 import { ModalPicker, NestedPicker } from './layout/'
 import { defaultProps } from 'recompose'
-
+import { defaultTheme } from './utils/hoc'
+/*
 export default ({
   theme,
   label = 'Add Custom Filter',
@@ -11,7 +13,15 @@ export default ({
       theme: {
         ...theme,
         Picker: defaultProps({ theme })(theme && theme.Picker || NestedPicker),
-    },
-    label,
-  })(ModalPicker),
+      },
+      label,
+    })(ModalPicker),
   })(FilterAdder)
+*/
+
+let DefaultPicker = defaultTheme({ Picker: NestedPicker })(ModalPicker)
+
+export default _.compose(
+  defaultProps({ label: 'Add Custom Filter' }),
+  defaultTheme({ Picker: DefaultPicker })
+)(FilterAdder)
