+++ src/exampleTypes/DateHistogram.js
@@ -1,16 +1,22 @@
 import React from 'react'
-import { contexturify } from '../utils/hoc'
+import _ from 'lodash/fp'
+import { contexturify, defaultTheme } from '../utils/hoc'
 import BarChart from '../layout/BarChart'
 
-let DateHistogram = contexturify(({ node, ...props }) => (
-  <BarChart
-    data={node.context.entries}
-    categoryField="key"
-    valueField="count"
-    gutter={0}
-    {...props}
-  />
-))
+let DateHistogram = _.flow(
+  defaultTheme({ BarChart }),
+  contexturify
+)(
+  ({ node, ...props }) => (
+    <theme.BarChart
+      data={node.context.entries}
+      categoryField="key"
+      valueField="count"
+      gutter={0}
+      {...props}
+    />
+  )
+)
 DateHistogram.displayName = 'DateHistogram'
 
 export default DateHistogram
