+++ src/exampleTypes/Text.js
@@ -1,9 +1,10 @@
 import _ from 'lodash/fp'
-import { contexturify, withTreeLens } from '../utils/hoc'
+import { contexturify, defaultTheme, withTreeLens } from '../utils/hoc'
 import LensInput from '../layout/LensInput'
 
 let Text = _.flow(
   withTreeLens,
+  defaultTheme({ Input: 'input' }),
   contexturify
 )(LensInput)
 Text.displayName = 'Text'
