+++ src/layout/Checkbox.js
@@ -1,0 +1,4 @@
+import React from 'react'
+
+let Checkbox = props => <input type="checkbox" {...props} />
+export default Checkbox
