+++ src/layout/LensInput.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { observer } from 'mobx-react'
 import F from 'futil-js'
 
-let LensInput = observer(({ lens, theme, ...x }) => (
-  <theme.Input {...F.domLens.value(lens)} {...x} />
+let LensInput = observer(({ lens, theme, ...props }) => (
+  <theme.Input {...F.domLens.value(lens)} {...props} />
 ))
 LensInput.displayName = 'LensInput'
