+++ src/layout/NumberInput.js
@@ -1,0 +1,4 @@
+import React from 'react'
+
+let NumberInput = props => <input type="number" {...props} />
+export default NumberInput
