+++ src/layout/index.js
@@ -1,39 +1,41 @@
+import { ModalPicker } from './Pickers'
+import { Tag, TagsInput } from './TagsInput'
 import Awaiter from './Awaiter'
+import BarChart from './BarChart'
+import CheckButton from './CheckButton'
+import Dynamic from './Dynamic'
+import ExpandableTable from './ExpandableTable'
 import { Flex } from './Flex'
 import { Grid } from './Grid'
-import Modal from './Modal'
-import Popover from './Popover'
-import SpacedList from './SpacedList'
-import TextHighlight from './TextHighlight'
-import { ModalPicker } from './Pickers'
-import NestedPicker from './NestedPicker'
-import Dynamic from './Dynamic'
-import BarChart from './BarChart'
-import { Tag, TagsInput } from './TagsInput'
 import LensInput from './LensInput'
-import ExpandableTable from './ExpandableTable'
+import Modal from './Modal'
+import NestedPicker from './NestedPicker'
+import NumberInput from './NumberInput'
+import Popover from './Popover'
 import Portal from './Portal'
-import CheckButton from './CheckButton'
+import SpacedList from './SpacedList'
 import StepsAccordion, { AccordionStep } from './StepsAccordion'
+import TextHighlight from './TextHighlight'
 
 export {
+  AccordionStep,
   Awaiter,
+  BarChart,
+  CheckButton,
+  Dynamic,
+  ExpandableTable,
   Flex,
   Grid,
+  LensInput,
   Modal,
-  Popover,
-  SpacedList,
-  TextHighlight,
-  NestedPicker,
   ModalPicker,
-  Dynamic,
-  BarChart,
+  NestedPicker,
+  NumberInput,
+  Popover,
+  Portal,
+  SpacedList,
+  StepsAccordion,
   Tag,
   TagsInput,
-  LensInput,
-  ExpandableTable,
-  Portal,
-  CheckButton,
-  StepsAccordion,
-  AccordionStep,
+  TextHighlight,
 }
